chore: test 시 profile 옵션 주입 기능 추가
