+++ build.gradle
@@ -71,3 +71,7 @@ tasks.named('bootBuildImage') {
 tasks.named('test') {
     useJUnitPlatform()
 }
+
+tasks.withType(Test).configureEach {
+    systemProperty 'spring.profiles.active', project.hasProperty('profile') ? project.profile : 'local'
+}
