+++ src/main/resources/application-dev.yml
@@ -11,8 +11,8 @@ spring:
 
 logging:
   file:
-    path: ${user.dir}/log
-    name: ${user.dir}/log/formssafe
+    path: ${user.dir}/../log
+    name: ${user.dir}/../log/formssafe
   pattern:
     file: "%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] [%X{traceId}] %highlight(%-5level) %cyan(%logger{36}.%M) - %msg%n"
   level:
