+++ src/test/resources/application-dev.yml
@@ -1,13 +1,11 @@
 spring:
   config:
     import: /home/ubuntu/web/privacy_test.yml
-    activate:
-      on-profile: dev
 
 logging:
   file:
-    path: ${user.dir}/log
-    name: ${user.dir}/log/formssafe
+    path: ${user.dir}/../log
+    name: ${user.dir}/../log/formssafe
   pattern:
     file: "%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] [%X{traceId}] %highlight(%-5level) %cyan(%logger{36}.%M) - %msg%n"
   level:
