+++ src/test/resources/application-local.yml
@@ -1,0 +1,3 @@
+spring:
+  config:
+    import: classpath:privacy.yml
