+++ src/test/resources/application.yml
@@ -1,5 +1,4 @@
 server:
-  port: 80
   servlet:
     session:
       cookie:
@@ -9,8 +8,6 @@ server:
       timeout: 3600
 
 spring:
-  config:
-    import: classpath:privacy.yml
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://${db.host}:${db.port}/${db.database}?useSSL=false&useUnicode=true&serverTimezone=Asia/Seoul&allowPublicKeyRetrieval=true
